experimental: extend builds and minor refactoring
- Adjusts the autogen.sh build to be more general (no need to have
  Makefile as this can be generated by the autogen.sh in some cases).
- Uses a unified place for templating
- Minor adjustment in directory name of shared folder
  "autogen-results-X" is now always just "autogen-results"
