+++ experimental/c-cpp/runner.py
@@ -22,63 +22,12 @@ import sys
 import threading
 from typing import List
 
+import templates
+
 silent_global = False
 
-empty_oss_fuzz_build = """#!/bin/bash -eu
-# Copyright 2018 Google Inc.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#      http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-#
-################################################################################
-"""
 
-empty_oss_fuzz_docker = """# Copyright 2018 Google Inc.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#      http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-#
-################################################################################
-
-FROM gcr.io/oss-fuzz-base/base-builder
-RUN apt-get update && apt-get install -y make autoconf automake libtool cmake \
-                      pkg-config curl check libcpputest-dev re2c
-RUN rm /usr/local/bin/cargo && \
- curl --proto '=https' --tlsv1.2 -sSf https://sh.rustup.rs | bash -s -- -y && \
- apt-get install -y cargo
-RUN python3 -m pip install --upgrade pip && \
-    python3 -m pip install pydantic-core pyyaml cxxfilt openai==1.16.2
-RUN python3 -m pip install --upgrade google-cloud-aiplatform
-COPY *.py *.json $SRC/
-WORKDIR $SRC
-COPY build.sh $SRC/"""
-
-empty_project_yaml = """homepage: "https://github.com/google/oss-fuzz"
-language: c++
-primary_contact: "info@oss-fuzz.com"
-auto_ccs:
--
-main_repo: 'https://github.com/samtools/htslib.git'
-"""
-
+SHARED_MEMORY_RESULTS_DIR='autogen-results'
 
 def setup_worker_project(oss_fuzz_base: str, project_name: str, llm_model: str):
   """Setup empty OSS-Fuzz project used for analysis."""
@@ -88,11 +37,11 @@ def setup_worker_project(oss_fuzz_base: str, project_name: str, llm_model: str):
 
   os.makedirs(temp_project_dir)
   with open(os.path.join(temp_project_dir, 'project.yaml'), 'w') as f:
-    f.write(empty_project_yaml)
+    f.write(templates.empty_project_yaml)
   with open(os.path.join(temp_project_dir, 'build.sh'), 'w') as f:
-    f.write(empty_oss_fuzz_build)
+    f.write(templates.empty_oss_fuzz_build)
   with open(os.path.join(temp_project_dir, 'Dockerfile'), 'w') as f:
-    f.write(empty_oss_fuzz_docker)
+    f.write(templates.AUTOGEN_DOCKER_FILE)
 
   if llm_model == 'vertex':
     json_config = os.environ.get('GOOGLE_APPLICATION_CREDENTIALS', None)
@@ -103,7 +52,7 @@ def setup_worker_project(oss_fuzz_base: str, project_name: str, llm_model: str):
     shutil.copyfile(json_config, os.path.join(temp_project_dir, 'creds.json'))
 
   # Copy over the generator
-  files_to_copy = {'build_generator.py', 'manager.py'}
+  files_to_copy = {'build_generator.py', 'manager.py', 'templates.py'}
   for target_file in files_to_copy:
     shutil.copyfile(
         os.path.join(os.path.dirname(os.path.abspath(__file__)), target_file),
@@ -132,7 +81,7 @@ def run_coverage_runs(oss_fuzz_base: str, worker_name: str) -> None:
   the generation is based on the OSS-Fuzz base-builder image."""
   worker_out = os.path.join(
       oss_fuzz_base, 'build', 'out', worker_name,
-      'autogen-results-%d' % (int(worker_name.split('-')[-1])))
+      SHARED_MEMORY_RESULTS_DIR)
 
   for auto_fuzz_dir in os.listdir(worker_out):
     print(auto_fuzz_dir)
@@ -277,7 +226,7 @@ def run_on_targets(target,
 
   openai_api_key = os.getenv('OPENAI_API_KEY', None)
 
-  outdir = '/out/autogen-results-%d' % (idx)
+  outdir = '/out/' + SHARED_MEMORY_RESULTS_DIR
   with open('status-log.txt', 'a') as f:
     f.write("Targeting: %s :: %d\n" % (target, idx))
   run_autogen(target,
