+++ experimental/c-cpp/templates.py
@@ -1,0 +1,158 @@
+#!/usr/bin/env python3
+# Copyright 2024 Google LLC
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+empty_oss_fuzz_build = '''#!/bin/bash -eu
+# Copyright 2024 Google LLC.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#      http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
+################################################################################
+'''
+
+base_docker_file = '''# Copyright 2018 Google Inc.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#      http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
+################################################################################
+
+FROM gcr.io/oss-fuzz-base/base-builder
+RUN apt-get update && apt-get install -y make autoconf automake autopoint \
+                      libtool cmake pkg-config curl check libcpputest-dev \
+                      flex bison re2c
+'''
+
+CFLITE_TEMPLATE = '''name: ClusterFuzzLite PR fuzzing
+on:
+  workflow_dispatch:
+  pull_request:
+    branches: [ master ]
+permissions: read-all
+jobs:
+  PR:
+    runs-on: ubuntu-latest
+    strategy:
+      fail-fast: false
+      matrix:
+        sanitizer: [address]
+    steps:
+    - name: Build Fuzzers (${{ matrix.sanitizer }})
+      id: build
+      uses: google/clusterfuzzlite/actions/build_fuzzers@v1
+      with:
+        sanitizer: ${{ matrix.sanitizer }}
+        language: c++
+        bad-build-check: false
+    - name: Run Fuzzers (${{ matrix.sanitizer }})
+      id: run
+      uses: google/clusterfuzzlite/actions/run_fuzzers@v1
+      with:
+        github-token: ${{ secrets.GITHUB_TOKEN }}
+        fuzz-seconds: 100
+        mode: 'code-change'
+        report-unreproducible-crashes: false
+        sanitizer: ${{ matrix.sanitizer }}
+'''
+
+# Empty CPP harness that is used to confirm compilation when generating
+# auto-build scripts.
+CPP_BASE_TEMPLATE = '''#include <stdint.h>
+#include <iostream>
+
+extern "C" int
+LLVMFuzzerTestOneInput(const uint8_t *data, size_t size)
+{
+    std::string input(reinterpret_cast<const char*>(data), size);
+
+    // Insert fuzzer contents here
+    // input string contains fuzz input.
+
+    // end of fuzzer contents
+
+    return 0;
+}'''
+
+# Empty C harness that is used to confirm compilation when generating
+# auto-build scripts.
+C_BASE_TEMPLATE = '''#include <stdint.h>
+#include <stdlib.h>
+#include <stdio.h>
+
+int LLVMFuzzerTestOneInput(const uint8_t *data, size_t size)
+{
+    printf("Hello world\\n");
+    // Insert fuzzer contents here
+    // input string contains fuzz input.
+
+    // end of fuzzer contents
+
+    return 0;
+}'''
+
+
+# Docker file used for starting the auto-gen workflow within an OSS-Fuzz
+# base-builder image.
+AUTOGEN_DOCKER_FILE = base_docker_file + '''
+RUN rm /usr/local/bin/cargo && \
+ curl --proto '=https' --tlsv1.2 -sSf https://sh.rustup.rs | bash -s -- -y && \
+ apt-get install -y cargo
+RUN python3 -m pip install --upgrade pip && \
+    python3 -m pip install pydantic-core pyyaml cxxfilt openai==1.16.2
+RUN python3 -m pip install --upgrade google-cloud-aiplatform
+COPY *.py *.json $SRC/
+WORKDIR $SRC
+COPY build.sh $SRC/'''
+
+empty_project_yaml = """homepage: "https://github.com/google/oss-fuzz"
+language: c++
+primary_contact: "info@oss-fuzz.com"
+auto_ccs:
+-
+main_repo: 'https://github.com/google/oss-fuzz'
+"""
+
+# Docker file used for OSS-Fuzz integrations.
+CLEAN_OSS_FUZZ_DOCKER = base_docker_file + '''
+COPY *.sh *.cpp *.c $SRC/
+RUN git clone --recurse-submodules {repo_url} {project_repo_dir}
+WORKDIR {project_repo_dir}
+'''
+
+CLEAN_DOCKER_CFLITE = base_docker_file + '''
+COPY . $SRC/{project_repo_dir}
+COPY .clusterfuzzlite/build.sh $SRC/build.sh
+COPY .clusterfuzzlite/*.cpp $SRC/
+COPY .clusterfuzzlite/*.c $SRC/
+WORKDIR {project_repo_dir}
+'''
